requirements.txt
requests==2.31.0
twilio==9.0.5
python-dotenv==1.0.1